secure-chat-app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.0
python-dotenv==1.0.0
websockets==12.0
aiosqlite==0.19.0
python-multipart==0.0.6
